+++ minimalkv/memory/redisstore.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 import re
 from io import BytesIO
-from typing import IO, TYPE_CHECKING, Iterator, List, Optional, Union
+from typing import TYPE_CHECKING, BinaryIO, Iterator, List, Optional, Union
 
 if TYPE_CHECKING:
     from redis import StrictRedis
@@ -63,11 +63,11 @@ class RedisStore(TimeToLiveMixin, KeyValueStore):
             raise KeyError(key)
         return val
 
-    def _get_file(self, key: str, file: IO) -> str:
+    def _get_file(self, key: str, file: BinaryIO) -> str:
         file.write(self._get(key))
         return key
 
-    def _open(self, key: str) -> IO:
+    def _open(self, key: str) -> BinaryIO:
         return BytesIO(self._get(key))
 
     def _put(
@@ -95,7 +95,10 @@ class RedisStore(TimeToLiveMixin, KeyValueStore):
         return key
 
     def _put_file(
-        self, key: str, file: IO, ttl_secs: Optional[Union[str, int, float]] = None
+        self,
+        key: str,
+        file: BinaryIO,
+        ttl_secs: Optional[Union[str, int, float]] = None,
     ) -> str:
         self._put(key, file.read(), ttl_secs)
         return key