+++ docs/changes.rst
@@ -1,6 +1,10 @@
 Changelog
 *********
 
+1.8.5
+=====
+* Changed generic `IO` type to `BinaryIO`.
+
 1.8.4
 =====
 * Removing invalid BSD-3 Clause license classifier.